+++ remember-me-controls.php
@@ -45,7 +45,7 @@ if ( ! class_exists( 'c2c_RememberMeControls' ) ) :
 
 require_once( dirname( __FILE__ ) . DIRECTORY_SEPARATOR . 'c2c-plugin.php' );
 
-final class c2c_RememberMeControls extends c2c_Plugin_064 {
+final class c2c_RememberMeControls extends c2c_Plugin_065 {
 
 	/**
 	 * Name of plugin's setting.