+++ tests/phpunit/tests/remember-me-controls.php
@@ -105,11 +105,11 @@ JS;
 	}
 
 	public function test_plugin_framework_class_name() {
-		$this->assertTrue( class_exists( 'c2c_Plugin_064' ) );
+		$this->assertTrue( class_exists( 'c2c_Plugin_065' ) );
 	}
 
 	public function test_plugin_framework_version() {
-		$this->assertEquals( '064', $this->obj->c2c_plugin_version() );
+		$this->assertEquals( '065', $this->obj->c2c_plugin_version() );
 	}
 
 	public function test_get_version() {